# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SD Department Enrollment Queue Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /counters:
    get:
      operationId: listCounters
      tags: [counters]
      summary: List all counters (evaluators and taggers)
      responses:
        "200":
          description: List of counters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Counter"
    post:
      operationId: createCounter
      tags: [counters]
      summary: Create a new counter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterInput"
      responses:
        "201":
          description: Counter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
  /counters/{id}:
    get:
      operationId: getCounter
      tags: [counters]
      summary: Get a counter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Counter details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
        "404":
          description: Counter not found
    patch:
      operationId: updateCounter
      tags: [counters]
      summary: Update counter status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterUpdate"
      responses:
        "200":
          description: Counter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
        "404":
          description: Counter not found
    delete:
      operationId: deleteCounter
      tags: [counters]
      summary: Delete a counter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Counter deleted
  /staff/login:
    post:
      operationId: staffLogin
      tags: [counters]
      summary: Staff login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffLoginResult"
        "401":
          description: Invalid credentials
  /counters/{id}/login:
    post:
      operationId: counterLogin
      tags: [counters]
      summary: Staff login to a counter with PIN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
        "401":
          description: Invalid PIN
        "404":
          description: Counter not found

  # Queue management
...
components:
  schemas:
...
    StaffLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    StaffLoginResult:
      type: object
      required: [token, role, counterId, counterName]
      properties:
        token:
          type: string
        role:
          type: string
          enum: [evaluator, tagger, hybrid, admin]
        counterId:
          type: integer
        counterName:
          type: string
    Counter:
      type: object
      required: [id, name, type, status, isActive, isOnline, currentQueueCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [evaluator, tagger, hybrid]
        specialization:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, busy, paused, break, offline]
        isActive:
          type: boolean
        isOnline:
          type: boolean
        currentQueueCount:
          type: integer
        currentStudentId:
          type: ["integer", "null"]
        currentStudentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CounterInput:
      type: object
      required: [name, type, username, pin]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [evaluator, tagger, hybrid]
        specialization:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        username:
          type: string
          minLength: 3
        pin:
          type: string
          minLength: 4
    CounterUpdate:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        username:
          type: string
          minLength: 3
        pin:
          type: string
          minLength: 4
        status:
          type: string
          enum: [available, busy, paused, break, offline]
        isActive:
          type: boolean
    CounterLoginInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
...