# lib/api-zod/src/generated/types/staffLoginResult.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * SD Department Enrollment Queue Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffLoginResultRole } from './staffLoginResultRole';

export interface StaffLoginResult {
  token: string;
  role: StaffLoginResultRole;
  counterId: number;
  counterName: string;
}